+++ apps/web/src/queries/saveSongQuery.ts
@@ -1,13 +1,6 @@
 import { useMutation, useQuery, useQueryClient } from '@tanstack/react-query';
 
 import { deleteSaveSong, getSaveSong, patchSaveSong } from '@/lib/api/saveSong';
-import {
-  deleteSaveFolderSong,
-  getSaveFolderSong,
-  patchSaveFolderSong,
-  postSaveFolderSong,
-} from '@/lib/api/saveSongFolder';
-import { postTotalStat, postTotalStatArray } from '@/lib/api/totalStat';
 import { SaveSong, SaveSongFolder } from '@/types/song';
 
 export function useSaveSongQuery() {
@@ -90,85 +83,6 @@ export function useDeleteSaveSongMutation() {
     },
     onSettled: () => {
       queryClient.invalidateQueries({ queryKey: ['saveSongFolder'] });
-    },
-  });
-}
-
-export function useSaveSongFolderQuery() {
-  return useQuery({
-    queryKey: ['saveSongFolderList'],
-    queryFn: async () => {
-      const response = await getSaveFolderSong();
-      if (!response.success || !response.data) {
-        return [];
-      }
-
-      return response.data;
-    },
-    staleTime: 1000 * 60 * 5,
-    gcTime: 1000 * 60 * 5,
-  });
-}
-
-export function usePostSaveSongFolderMutation() {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    mutationFn: async ({ folderName }: { folderName: string }) => {
-      const data = await postSaveFolderSong({ folderName });
-      console.log('usePostSaveSongFolderMutation', data);
-      if (!data.success) {
-        throw new Error(data.error);
-      }
-    },
-    onError: error => {
-      console.log('error', error);
-      alert(error.message ?? 'POST 실패');
-    },
-    onSettled: () => {
-      queryClient.invalidateQueries({ queryKey: ['saveSongFolderList'] });
-    },
-  });
-}
-
-export function useRenameSaveSongFolderMutation() {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    mutationFn: async ({ folderId, folderName }: { folderId: string; folderName: string }) => {
-      const data = await patchSaveFolderSong({ folderId, folderName });
-      console.log('useRenameSaveSongFolderMutation', data);
-      if (!data.success) {
-        throw new Error(data.error);
-      }
-    },
-    onError: error => {
-      console.log('error', error);
-      alert(error.message ?? 'POST 실패');
-    },
-    onSettled: () => {
-      queryClient.invalidateQueries({ queryKey: ['saveSongFolderList'] });
-    },
-  });
-}
-
-export function useDeleteSaveFolderSongMutation() {
-  const queryClient = useQueryClient();
-
-  return useMutation({
-    mutationFn: async ({ folderId }: { folderId: string }) => {
-      const data = await deleteSaveFolderSong({ folderId });
-      console.log('useDeleteSaveFolderSongMutation', data);
-      if (!data.success) {
-        throw new Error(data.error);
-      }
-    },
-    onError: error => {
-      console.log('error', error);
-      alert(error.message ?? 'POST 실패');
-    },
-    onSettled: () => {
-      queryClient.invalidateQueries({ queryKey: ['saveSongFolderList'] });
     },
   });
 }
